mat4 development

+++ src/core/math/vec3.hpp
@@ -3,7 +3,6 @@
 #include<iostream>
 #include<iomanip>
 #include<cassert>
-#include<cstdint>
 
 #include"simd_backend.hpp"
 #include"vec3packed.hpp"
@@ -86,12 +85,12 @@ struct alignas(16) Vec3{
 		return *this;
 	}
 
-	FORCE_INLINE const float& operator[](uint16_t i) const {
+	FORCE_INLINE const float& operator[](int i) const {
 		assert(i < 3 && "index oob for Vec3");
 		return (&x)[i];
 	}
 
-	FORCE_INLINE float& operator[](uint16_t i) {
+	FORCE_INLINE float& operator[](int i) {
 		assert(i < 3 && "index oob for Vec3");
 		return (&x)[i];
 	}
